Organized Orgs/Projects/Tech
Part of fix for issue #9
This makes it easier to build modes/views/etc
