+++ Portfolio/Orgs/Data/OrganizationTable.cs
@@ -1,26 +1,32 @@
 using System.Data;
 using System.Data.SQLite;
 using ExperimentalSQLite;
+using Portfolio.Projects.Data;
 using WebServer.Models;
 
-namespace Portfolio.Projects {
-    public class OrganizationTable : PortfolioDatabase.SQLiteTable<OrganizationTable, OrganizationInfo> {
+namespace Portfolio.Orgs.Data
+{
+    public class OrganizationTable : PortfolioDatabase.SQLiteTable<OrganizationTable, OrganizationInfo>
+    {
         public OrganizationTable(PortfolioDatabase database) : base(database, nameof(OrganizationInfo)) { }
 
         public override OrganizationInfo ConstructRow() => new OrganizationInfo(this);
 
-        public IEnumerable<OrganizationInfo> GetOrgInfoFromProjectInfo(IEnumerable<ProjectInfo> projInfo, bool ensureOrgPublised) {
+        public IEnumerable<OrganizationInfo> GetOrgInfoFromProjectInfo(IEnumerable<ProjectInfo> projInfo, bool ensureOrgPublised)
+        {
             // Since the values are all ints, I won't use SQL parameters here
             if (!projInfo.Any()) return Array.Empty<OrganizationInfo>();
             projInfo = projInfo.Where(projInfo => projInfo.OrganizationId.Value != null);
             string whereClause = $"({string.Join(" OR ", projInfo.Select(techItem => $"`{Schema.OrganizationId.ColumnName}` = {techItem.OrganizationId.Value}"))})" +
                 $" AND `{Schema.IsPublished.ColumnName}` = {true};";
-            using (SQLiteCommand cmd = new($"SELECT * FROM `{TableName}` WHERE {whereClause};", Database.Connection)) {
+            using (SQLiteCommand cmd = new($"SELECT * FROM `{TableName}` WHERE {whereClause};", Database.Connection))
+            {
                 return ReadFromReader(cmd.ExecuteReader());
             }
         }
     }
-    public class OrganizationInfo : OrganizationTable.SQLiteRow, IPageModel {
+    public class OrganizationInfo : OrganizationTable.SQLiteRow, IPageModel
+    {
         public override IEnumerable<IDbCell> Fields => new IDbCell[] { OrganizationId, UrlName, Name, Role, StartTimestamp, EndTimestamp, BriefText, OverviewSubheaderOverride, OverviewMarkdown, IsPublished };
         public Dictionary<string, object> Values => new(Fields.Omit(OrganizationId, IsPublished).ToDictionary(cell => cell.ColumnName, cell => cell.Value)) {
             { "Duration", StartTimestamp.Value.ToDurationString(EndTimestamp.Value) }
